requirements.txt
pyspark==3.5.1
kagglehub==0.2.7
python-dotenv==1.0.1
